Move all files from interface/src to plugins for TB simulation
