+++ SimG4CMS/HGCalTestBeam/plugins/BuildFile.xml
@@ -12,7 +12,6 @@
 <use   name="SimDataFormats/HcalTestBeam"/>
 <use   name="SimDataFormats/Track"/>
 <use   name="SimDataFormats/Vertex"/>
-<use   name="SimG4CMS/HGCalTestBeam"/>
 <use   name="SimG4CMS/Calo"/>
 <use   name="SimG4Core/Notification"/>
 <use   name="SimG4Core/Watcher"/>
