+++ SimG4CMS/HGCalTestBeam/plugins/HGCalTBAnalyzer.cc
@@ -35,8 +35,8 @@
 #include "SimDataFormats/HcalTestBeam/interface/HcalTestBeamNumbering.h"
 #include "SimDataFormats/Track/interface/SimTrackContainer.h"
 #include "SimDataFormats/Vertex/interface/SimVertexContainer.h"
-#include "SimG4CMS/HGCalTestBeam/interface/AHCalDetId.h"
-#include "SimG4CMS/HGCalTestBeam/interface/AHCalGeometry.h"
+#include "AHCalDetId.h"
+#include "AHCalGeometry.h"
 
 #include "SimDataFormats/CaloHit/interface/PassiveHit.h"
