--- SimG4CMS/HcalTestBeam/BuildFile.xml
@@ -1,21 +1,0 @@
-<use   name="SimG4Core/Notification"/>
-<use   name="SimG4Core/Watcher"/>
-<use   name="SimG4CMS/Calo"/>
-<use   name="Geometry/EcalCommonData"/>
-<use   name="Geometry/HcalCommonData"/>
-<use   name="DataFormats/HcalDetId"/>
-<use   name="DataFormats/Math"/>
-<use   name="SimDataFormats/HcalTestBeam"/>
-<use   name="FWCore/Framework"/>
-<use   name="FWCore/ParameterSet"/>
-<use   name="FWCore/MessageLogger"/>
-<use   name="FWCore/ServiceRegistry"/>
-<use   name="CommonTools/UtilAlgos"/>
-<use   name="geant4core"/>
-<use   name="root"/>
-<use   name="hepmc"/>
-<use   name="rootmath"/>
-<export>
-  <lib   name="1"/>
-</export>
-
