+++ SimG4CMS/HcalTestBeam/plugins/BuildFile.xml
@@ -1,10 +1,22 @@
-<use   name="SimG4CMS/HcalTestBeam"/>
-<use   name="SimG4Core/Watcher"/>
+<use   name="DataFormats/HcalDetId"/>
+<use   name="DataFormats/Math"/>
+<use   name="Geometry/EcalCommonData"/>
+<use   name="Geometry/HcalCommonData"/>
 <use   name="SimDataFormats/HcalTestBeam"/>
+<use   name="SimG4Core/Notification"/>
+<use   name="SimG4Core/Watcher"/>
+<use   name="SimG4CMS/Calo"/>
 <use   name="FWCore/Framework"/>
 <use   name="FWCore/ParameterSet"/>
 <use   name="FWCore/MessageLogger"/>
 <use   name="FWCore/ServiceRegistry"/>
+<use   name="CommonTools/UtilAlgos"/>
+<use   name="geant4core"/>
+<use   name="hepmc"/>
+<use   name="clhep"/>
+<use   name="rootmath"/>
+<use   name="rootphysics"/>
+<use   name="root"/>
 <flags   EDM_PLUGIN="1"/>
 <library   file="*.cc" name="SimG4CMSHcalTestBeamPlugins">
   <flags   EDM_PLUGIN="1"/>
