+++ SimG4CMS/HcalTestBeam/plugins/HcalTB02Analysis.cc
@@ -14,26 +14,40 @@
 #include <cmath>
 #include <iostream>
 #include <iomanip>
+#include <memory>
+#include <map>
+#include <vector>
+#include <string>
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB02Analysis.h"
-
 // to retreive hits
+#include "SimDataFormats/HcalTestBeam/interface/HcalTB02HistoClass.h"
 #include "SimG4CMS/Calo/interface/CaloG4Hit.h"
 #include "SimG4CMS/Calo/interface/CaloG4HitCollection.h"
 #include "SimG4CMS/HcalTestBeam/interface/HcalTB02HcalNumberingScheme.h"
 #include "SimG4CMS/HcalTestBeam/interface/HcalTB02XtalNumberingScheme.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Notification/interface/BeginOfJob.h"
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Watcher/interface/SimProducer.h"
 #include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "CLHEP/Random/RandGaussQ.h"
 
-#include "G4SDManager.hh"
-#include "G4VProcess.hh"
+#include "HcalTB02Histo.h"
+
 #include "G4HCofThisEvent.hh"
+#include "G4SDManager.hh"
+#include "G4Step.hh"
+#include "G4Track.hh"
+#include "G4ThreeVector.hh"
+#include "G4VProcess.hh"
+
+#include "CLHEP/Random/RandGaussQ.h"
 #include "CLHEP/Units/GlobalSystemOfUnits.h"
 #include "CLHEP/Units/GlobalPhysicalConstants.h"
 #include "Randomize.hh"
@@ -41,6 +55,50 @@
 namespace CLHEP {
   class HepRandomEngine;
 }
+
+class HcalTB02Analysis : public SimProducer,
+                         public Observer<const BeginOfEvent *>,
+                         public Observer<const EndOfEvent *> {
+public:
+  HcalTB02Analysis(const edm::ParameterSet &p);
+  ~HcalTB02Analysis() override;
+
+  void produce(edm::Event &, const edm::EventSetup &) override;
+
+private:
+  HcalTB02Analysis(const HcalTB02Analysis &) = delete;  // stop default
+  const HcalTB02Analysis &operator=(const HcalTB02Analysis &) = delete;
+
+  // observer methods
+  void update(const BeginOfEvent *evt) override;
+  void update(const EndOfEvent *evt) override;
+
+  void fillEvent(HcalTB02HistoClass &);
+  void clear();
+  void finish();
+
+private:
+  // Private Tuples
+  HcalTB02Histo *histo;
+
+  // to read from parameter set
+  bool hcalOnly;
+  std::string fileNameTuple;
+  std::vector<std::string> names;
+
+  //To be saved
+  std::map<int, float> energyInScints, energyInCrystals;
+  std::map<int, float> primaries;
+  int particleType;
+  double eta, phi, pInit, incidentEnergy;
+  float SEnergy, E7x7Matrix, E5x5Matrix;
+  float SEnergyN, E7x7MatrixN, E5x5MatrixN;
+  int maxTime;
+  double xIncidentEnergy;
+  float xSEnergy, xSEnergyN;
+  float xE3x3Matrix, xE5x5Matrix;
+  float xE3x3MatrixN, xE5x5MatrixN;
+};
 
 //
 // constructors and destructor
