+++ SimG4CMS/HcalTestBeam/plugins/HcalTB02HcalNumberingScheme.cc
@@ -13,12 +13,10 @@
 // system include files
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB02HcalNumberingScheme.h"
+#include "HcalTB02HcalNumberingScheme.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
 #include "G4SystemOfUnits.hh"
-
-using namespace std;
 
 //
 // constructors and destructor
@@ -26,11 +24,11 @@ using namespace std;
 
 HcalTB02HcalNumberingScheme::HcalTB02HcalNumberingScheme()
     : HcalTB02NumberingScheme(), phiScale(1000000), etaScale(10000) {
-  edm::LogInfo("HcalTBSim") << "Creating HcalTB02HcalNumberingScheme";
+  edm::LogVerbatim("HcalTBSim") << "Creating HcalTB02HcalNumberingScheme";
 }
 
 HcalTB02HcalNumberingScheme::~HcalTB02HcalNumberingScheme() {
-  edm::LogInfo("HcalTBSim") << "Deleting HcalTB02HcalNumberingScheme";
+  edm::LogVerbatim("HcalTBSim") << "Deleting HcalTB02HcalNumberingScheme";
 }
 
 //
@@ -57,17 +55,17 @@ int HcalTB02HcalNumberingScheme::getUnitID(const G4Step* aStep) const {
   // Compute the scintID in the HB.
 
   float hR = hitPoint.mag();  //sqrt( pow(hx,2)+pow(hy,2)+pow(hz,2) );
-  float htheta = (hR == 0. ? 0. : acos(max(min(hz / hR, float(1.)), float(-1.))));
+  float htheta = (hR == 0. ? 0. : acos(std::max(std::min(hz / hR, float(1.)), float(-1.))));
   float hsintheta = sin(htheta);
-  float hphi = (hR * hsintheta == 0. ? 0. : acos(max(min(hx / (hR * hsintheta), float(1.)), float(-1.))));
+  float hphi = (hR * hsintheta == 0. ? 0. : acos(std::max(std::min(hx / (hR * hsintheta), float(1.)), float(-1.))));
   float heta = (fabs(hsintheta) == 1. ? 0. : -log(fabs(tan(htheta / 2.))));
   int eta = int(heta / 0.087);
   int phi = int(hphi / (5. * degree));
 
   G4VPhysicalVolume* thePV = preStepPoint->GetPhysicalVolume();
   int ilayer = ((thePV->GetCopyNo()) / 10) % 100;
-  LogDebug("HcalTBSim") << "HcalTB02HcalNumberingScheme:: Layer " << thePV->GetName() << " found at phi = " << phi
-                        << " eta = " << eta << " lay = " << thePV->GetCopyNo() << " " << ilayer;
+  edm::LogVerbatim("HcalTBSim") << "HcalTB02HcalNumberingScheme:: Layer " << thePV->GetName() << " found at phi = " 
+				<< phi << " eta = " << eta << " lay = " << thePV->GetCopyNo() << " " << ilayer;
 
   scintID = phiScale * phi + etaScale * eta + ilayer;
   if (hy < 0.)
@@ -82,7 +80,7 @@ int HcalTB02HcalNumberingScheme::getlayerID(int sID) const {
   if ((layerID != 17) && (layerID != 18))
     layerID = sID - int(float(sID) / float(etaScale)) * etaScale;
 
-  LogDebug("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " layer = " << layerID;
+  edm::LogVerbatim("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " layer = " << layerID;
   return layerID;
 }
 
@@ -92,7 +90,7 @@ int HcalTB02HcalNumberingScheme::getphiID(int sID) const {
     IDsign = -1;
   sID = abs(sID);
   int phiID = int(float(sID) / float(phiScale));
-  LogDebug("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " phi = " << phiID;
+  edm::LogVerbatim("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " phi = " << phiID;
   if (IDsign > 0) {
     phiID += 4;
   } else {
@@ -106,6 +104,6 @@ int HcalTB02HcalNumberingScheme::getetaID(int sID) const {
   int aux = sID - int(float(sID) / float(phiScale)) * phiScale;
   int etaID = int(float(aux) / float(etaScale));
 
-  LogDebug("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " eta = " << etaID;
+  edm::LogVerbatim("HcalTBSim") << "HcalTB02HcalNumberingScheme:: scintID " << sID << " eta = " << etaID;
   return etaID;
 }
