+++ SimG4CMS/HcalTestBeam/plugins/HcalTB02Histo.cc
@@ -15,10 +15,10 @@
 #include <cmath>
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB02Histo.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "HcalTB02Histo.h"
 
 //
 // constructors and destructor
