+++ SimG4CMS/HcalTestBeam/plugins/HcalTB02XtalNumberingScheme.cc
@@ -13,7 +13,7 @@
 // system include files
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB02XtalNumberingScheme.h"
+#include "HcalTB02XtalNumberingScheme.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
 //
@@ -21,11 +21,11 @@
 //
 
 HcalTB02XtalNumberingScheme::HcalTB02XtalNumberingScheme() : HcalTB02NumberingScheme() {
-  edm::LogInfo("HcalTBSim") << "Creating HcalTB02XtalNumberingScheme";
+  edm::LogVerbatim("HcalTBSim") << "Creating HcalTB02XtalNumberingScheme";
 }
 
 HcalTB02XtalNumberingScheme::~HcalTB02XtalNumberingScheme() {
-  edm::LogInfo("HcalTBSim") << "Deleting HcalTB02XtalNumberingScheme";
+  edm::LogVerbatim("HcalTBSim") << "Deleting HcalTB02XtalNumberingScheme";
 }
 
 //
@@ -39,6 +39,6 @@ int HcalTB02XtalNumberingScheme::getUnitID(const G4Step* aStep) const {
   if (touch->GetHistoryDepth() > 0)
     idl = touch->GetReplicaNumber(1);
   int idunit = idl * 100 + idx;
-  LogDebug("HcalTBSim") << "HcalTB02XtalNumberingScheme:: Row " << idl << " Column " << idl << " idunit = " << idunit;
+  edm::LogVerbatim("HcalTBSim") << "HcalTB02XtalNumberingScheme:: Row " << idl << " Column " << idl << " idunit = " << idunit;
   return idunit;
 }
