+++ SimG4CMS/HcalTestBeam/plugins/HcalTB04Analysis.cc
@@ -6,32 +6,38 @@
 // Implementation:
 //     Main analysis class for Hcal Test Beam 2004 Analysis
 //
+//  Usage: A Simwatcher class and can be activated from Oscarproducer module
+//
 // Original Author:
 //         Created:  Tue May 16 10:14:34 CEST 2006
 //
 
 // system include files
 #include <cmath>
-#include <iostream>
 #include <iomanip>
+#include <iostream>
+#include <memory>
+#include <vector>
+#include <string>
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB04Analysis.h"
-
+#include "SimG4Core/Notification/interface/Observer.h"
 #include "SimG4Core/Notification/interface/BeginOfRun.h"
 #include "SimG4Core/Notification/interface/BeginOfEvent.h"
 #include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Watcher/interface/SimProducer.h"
 #include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
 
 // to retreive hits
 #include "DataFormats/HcalDetId/interface/HcalSubdetector.h"
+#include "SimDataFormats/CaloHit/interface/CaloHit.h"
+#include "SimDataFormats/HcalTestBeam/interface/PHcalTB04Info.h"
 #include "SimG4CMS/Calo/interface/ECalSD.h"
 #include "SimG4CMS/Calo/interface/HCalSD.h"
+#include "SimG4CMS/Calo/interface/HcalQie.h"
 #include "SimG4CMS/Calo/interface/CaloG4Hit.h"
 #include "SimG4CMS/Calo/interface/CaloG4HitCollection.h"
 #include "SimG4CMS/Calo/interface/HcalTestNumberingScheme.h"
-#include "SimG4CMS/HcalTestBeam/interface/HcalTBNumberingScheme.h"
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB04XtalNumberingScheme.h"
 #include "DataFormats/Math/interface/Point3D.h"
 
 #include "FWCore/Framework/interface/Event.h"
@@ -41,14 +47,97 @@
 #include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
-#include "CLHEP/Random/RandGaussQ.h"
+#include "HcalTBNumberingScheme.h"
+#include "HcalTB04Histo.h"
+#include "HcalTB04XtalNumberingScheme.h"
 
 #include "G4SDManager.hh"
+#include "G4Step.hh"
+#include "G4Track.hh"
+#include "G4ThreeVector.hh"
 #include "G4VProcess.hh"
 #include "G4HCofThisEvent.hh"
+#include "CLHEP/Random/RandGaussQ.h"
 #include "CLHEP/Units/GlobalSystemOfUnits.h"
 #include "CLHEP/Units/GlobalPhysicalConstants.h"
 #include "Randomize.hh"
+#include <boost/cstdint.hpp>
+
+namespace CLHEP {
+  class HepRandomEngine;
+}
+
+class HcalTB04Analysis : public SimProducer,
+                         public Observer<const BeginOfRun *>,
+                         public Observer<const BeginOfEvent *>,
+                         public Observer<const EndOfEvent *>,
+                         public Observer<const G4Step *> {
+public:
+  HcalTB04Analysis(const edm::ParameterSet &p);
+  ~HcalTB04Analysis() override;
+
+  void produce(edm::Event &, const edm::EventSetup &) override;
+
+private:
+  HcalTB04Analysis(const HcalTB04Analysis &) = delete;  // stop default
+  const HcalTB04Analysis &operator=(const HcalTB04Analysis &) = delete;
+
+  void init();
+
+  // observer methods
+  void update(const BeginOfRun *run) override;
+  void update(const BeginOfEvent *evt) override;
+  void update(const G4Step *step) override;
+  void update(const EndOfEvent *evt) override;
+
+  //User methods
+  void fillBuffer(const EndOfEvent *evt);
+  void qieAnalysis(CLHEP::HepRandomEngine *);
+  void xtalAnalysis(CLHEP::HepRandomEngine *);
+  void finalAnalysis();
+  void fillEvent(PHcalTB04Info &);
+
+  void clear();
+  int unitID(uint32_t id);
+  double scale(int det, int layer);
+  double timeOfFlight(int det, int layer, double eta);
+
+private:
+  HcalQie *myQie;
+  HcalTB04Histo *histo;
+
+  // to read from parameter set
+  bool hcalOnly;
+  int mode, type;
+  double ecalNoise, beamOffset;
+  int iceta, icphi;
+  double scaleHB0, scaleHB16, scaleHO, scaleHE0;
+  std::vector<std::string> names;
+  G4RotationMatrix *beamline_RM;
+
+  // Constants for the run
+  int count;
+  int nTower, nCrystal;
+  std::vector<int> idHcal, idXtal;
+  std::vector<uint32_t> idTower, idEcal;
+
+  // Constants for the event
+  int nPrimary, particleType;
+  double pInit, etaInit, phiInit;
+  std::vector<CaloHit> ecalHitCache;
+  std::vector<CaloHit> hcalHitCache, hcalHitLayer;
+  std::vector<double> esimh, eqie, esime, enois;
+  std::vector<double> eseta, eqeta, esphi, eqphi, eslay, eqlay;
+  double etots, eecals, ehcals, etotq, eecalq, ehcalq;
+
+  bool pvFound;
+  int evNum, pvType;
+  G4ThreeVector pvPosition, pvMomentum, pvUVW;
+  std::vector<int> secTrackID, secPartID;
+  std::vector<G4ThreeVector> secMomentum;
+  std::vector<double> secEkin;
+  std::vector<int> shortLivedSecondaries;
+};
 
 //
 // constructors and destructor
