+++ SimG4CMS/HcalTestBeam/plugins/HcalTB06Histo.cc
@@ -15,10 +15,11 @@
 #include <cmath>
 
 // user include files
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB06Histo.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "CommonTools/UtilAlgos/interface/TFileService.h"
+
+#include "HcalTB06Histo.h"
 
 //
 // constructors and destructor
