+++ SimG4CMS/HcalTestBeam/plugins/module.cc
@@ -1,6 +1,5 @@
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB02SD.h"
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB06BeamSD.h"
-#include "SimG4CMS/HcalTestBeam/interface/HcalTB06Analysis.h"
+#include "HcalTB02SD.h"
+#include "HcalTB06BeamSD.h"
 #include "SimG4Core/SensitiveDetector/interface/SensitiveDetectorPluginFactory.h"
 #include "FWCore/PluginManager/interface/ModuleDef.h"
 
@@ -8,4 +7,3 @@ typedef HcalTB02SD HcalTB02SensitiveDetector;
 DEFINE_SENSITIVEDETECTOR(HcalTB02SensitiveDetector);
 typedef HcalTB06BeamSD HcalTB06BeamDetector;
 DEFINE_SENSITIVEDETECTOR(HcalTB06BeamDetector);
-DEFINE_FWK_MODULE(HcalTB06Analysis);
